Move GitHub helpers to scripts/utils/github/ and parameterize team membership
