+++ scripts/utils/github/constants.ts
@@ -1,0 +1,7 @@
+/**
+ * GitHub organization and repository identifiers shared by every script that
+ * talks to GitHub. Keep this file free of feature-specific constants; those
+ * belong with their feature.
+ */
+export const ORG = 'storybookjs';
+export const PRIMARY_REPO = 'storybook';
