+++ scripts/utils/github/linked-issues.test.ts
